Some changes
* Vector2 is now float-based
* took out enetity texture (tmp) as it was killing the program
* took out render offsets

+++ Goliath/Vector2.cpp
@@ -12,7 +12,7 @@ Vector2::Vector2()
 
 }
 
-Vector2::Vector2(int x, int y)
+Vector2::Vector2(float x, float y)
 {
     this->x = x;
     this->y = y;
